Major Refactor of the data model
- Changed the entire saving and DTO structure
- Moved images into the game.json
- Moved BgTiles and added UnitTiles to the game.json
- UnitMap and BgMap addd to game.json
- Setup gson serialization of Bitmaps
- Fixed code so it compiles

Warning!  Broken Version!  This will compile but has a lot of errors
We need a way to import images and tiles into a new game, or the games won't work
The changse to the storage mean that old games won't work any more, the variables
are all named differently and the storage is really moved around.